feat: add support for option useUnreleasedWithNoIncrement
This option serves the use case where:

- we are using the --no-increment flag
- we are on version `x.y.z` (you have bumped it through some other mechanism)
- we have no `x.y.z` section in our changelog file yet
- want want the changes from "Unreleased" to be used to create a the `x.y.z` section

This is useful in contexts where the version bump happens at a different
point in the development lifecycle from the updating of the changelog.
For example, when doing React Native development: the app store requires
all test flight builds to have a version number that is above the latest
released production version. This requires us to bump our version number
when we _start_ working on a new version, not when we release it.

So our workflow is:
- when we begin work on a new version, we bump the version (no updates to changelog)
- when our new version has been released to the app store, we use keep-a-changelog
to generate the release notes for that release